third_party/WebKit/LayoutTests/inspector/styles/stylesheet-source-url-comment-expected.txt

Tests that stylesheets with sourceURL comment are shown in the Sources panel.


Running: testSourceURLCommentInInlineScript

Running: testSourceURLComment
body { color: black; }
/*# sourceURL=css/addedInlineStylesheet.css */

Running: testDeprecatedSourceURLComment
body { color: black; }
/*@ sourceURL=css/addedInlineStylesheetDeprecated.css */

Running: testNonRelativeURL
body { color: red; }
/*# sourceURL=/css/nonRelativeInlineStylesheet.css */
